pyproject.toml
[tool.poetry]
name = "celebration-bot"
version = "0.1.0"
description = ""
authors = ["Leonard Henriquez <leonard@topo.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.32.1"
stripe = "^11.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
